Crawler WIP
It does not fall down, idk why, have a rough idea
You can jump over him and kill
Technically when you hit him from side or below you get pushed in
contrary direction, but instead it gets stuck for a time sometimes.
I have put a Timer to try and not do that but it ignores timer and
continues doing the collision effect.