Preparing Object Layer Loading
Want to load everything as Objects, helps with memory and how fast it will try and interact with collisions (way less colliders to interact with).

Also for entities, load just base stats and give them position to load to, instead of from the xml, which is messy and takes longer.